+++ tests/scenario/test_state.py
@@ -1,7 +1,4 @@
 import unittest
-import math
-
-import numpy as np
 
 from commonroad.scenario.state import *
 from commonroad.geometry.shape import Rectangle
@@ -167,6 +164,16 @@ class TestState(unittest.TestCase):
         self.assertEqual(custom_state_2.velocity.end, 1)
         self.assertEqual(custom_state_2.time_step.start, 0)
         self.assertEqual(custom_state_2.time_step.end, 1)
+
+    def test_extra_state_properties(self):
+        state = PMState(time_step=Interval(0, 1), velocity=0.0, velocity_y=0.0, position=np.array([0, 0]))
+        self.assertEqual(state.orientation, 0.0)
+        state = PMState(time_step=Interval(0, 1), velocity=0.0, velocity_y=1.0, position=np.array([0, 0]))
+        self.assertAlmostEqual(state.orientation, 1.57079, 3)
+
+        state = ExtendedPMState(time_step=Interval(0, 1), velocity=0.0, acceleration=1.0,
+                                position=np.array([0, 0]), orientation=0.0)
+        self.assertEqual(state.velocity_y, 0.0)
 
 
 if __name__ == '__main__':